iSCSILogicalUnit: do not use lio_iblock with lio-t
For lio-t, we cannot use the lio_iblock parameter, as targetcli just
ignores the index we give it and generates its own.

A logical consequence is that we cannot use the contents of the
lio_iblock parameter to access the iblock_* directories in the configfs,
since we do not know the values of the indices generated by targetcli.

This leads to an effect where systems with only one target work just
fine: targetcli always picks "0" as the first index, and the default
value of lio_iblock is "0". But when multiple targets are involved, one
of them will get the index "1" whereas we are still trying to access
"iblock_0", leading to errors of the form:

/sys/kernel/config/target/core/iblock_0/lu1_target1/wwn/product_id: No such file or directory

Since we only ever want to access the target directory inside the
iblock_* directory, we can just use a glob to sufficiently describe the
path -- there should always only be one match.