rabbitmq-server-ha: Avoid forget_cluster_node without node name
Currently when a rabbitmq-server resource is started,
the forget_cluster_node command is executed to ensure the node being
started is not part of the existing cluster.

This command is dependent on the local node name defined as a node
attribute but there are some situations like the examples below where
the attribute is not yet defined when the resource is started.
 - When a cluster is freshly deployed and started for the first time
 - When a cluster node is broken and replaced by a new node

This change ensures the node attribute exists to run the command,
to avoid invalid usage of the command without the target node name.