LVM-activate: adding new parameter auto_drop_lock
If VG is activated exclusively in the cluster and the lvmlockd daemon is
killed, the EX locks for LV will become orphaned in the lockspace. Then
LV will fail to be activated when cluster trying to bring it up
again. With this new parameter enabled, we drop the orphaned locks first
left in previous lockspace before activation.