azure-events-az: remove # from start of attr_healthstate attribute name
Prepending an attribute with # hides the value from "crm_mon -A" queries, but attribute is exposed when checking CIB.  Trying to add an attribute starting with # with "crm configure edit" results in an error in crmsh.  Suggesting this change to improve visibility and realign it with valid attribute naming convention.