fix podman-etcd: wrap ipv6 address in brackets for attribute_node_ip
When trying to determine the node ip address we need to make sure we account for ipv6 and dualstack deployments, and accordingly wrap ipv6 in brackets so it correctly resolves. We also need to prefer ipv6 over ipv4 in dualstack deployments, so when querying for the ip address, we favor ipv6.

Signed-off-by: ehila <ehila@redhat.com>