Refactor(podman-etcd): improve peer checking and leadership loss detection
The check_peers function is broken up into smaller, more manageable
functions. This refactoring separates the logic for detecting a loss of
cluster leadership from the logic for managing peer membership.

The main function is renamed to check_peer as there is only 1 peer to
check (it was check_peers).