Filesystem: speed up get_pids
With force_umount=safe, we "manually" scan the /proc/ file system.

We look for symlinks pointing into the path we are interested in.
Specifically, we are interested in
  /proc/<pid>/{root,exe,cwd}
  /proc/<pid>/fd/<fd>
We also look for relevant memory mappings in /proc/<pid>/maps

All these are per process, not per "task" or "thread".
see procfs(5) and pthreads(7).
Still, we currently also scan /proc/<pid>/task/<tid>/
for all the same things.

With a large system with many heavily threaded processes,
this can significantly slow down this scanning,
without gaining new information.

Adding -maxdepth to the find command line avoids this useless work,
potentially reducing the scanning time by orders of magnitute
on systems with many heavily threaded processes.

We could also write a dedicated helper in C to do the very same thing,
with the option to "short circuit" and proceed with the next pid
as soon as the first "match" is found for the currently inspected pid.

That could further reduce the scanning time
by about an additional factor of 10.