OCPEDGE-2231: podman-etcd: improve error handling to support retry on start errors (#2105)
* podman-etcd: improve add_member_as_learner error log

Improving add_member_as_learner error log to better debug rare issue
when the podman exec command returns error, but the etcd member is added
to the list anyway. This is critical as the `learner_node` attribute
won't be cleaned up anymore.

Signed-off-by: Carlo Lobrano <c.lobrano@gmail.com>

* podman-etcd: remove duplicated check for container already started

* podman-etcd: improve error return codes to support start retries

Improved and/or changed some returns code to allow or forbid retry in
case of start errors.

see: OCPEDGE-2231

---------

Signed-off-by: Carlo Lobrano <c.lobrano@gmail.com>