portblock: derive state file from parameters instead of instance
The expected usage of this agent is to pair a "block" with an "unblock",
and order startup and configuration of some service between these.

The established idiom is to have two separate instances with inverse actions.
To "reliably" report the status of "block" during a monitor action,
it is not sufficient to check the existence of the blocking rule.

It is also insufficient to rely on the pseudo resource state file
of this instance only.

To know our actual expectation, we need to check the state file of the
"inverse" instance as well.

Because we don't know the OCF_RESOURCE_INSTANCE value of the other instance,
we override the state file name for both instances to something derived from
our parameters.

This should give use the same "global state" view as the "promotion score"
does for the promotable clone variant of this agent.